% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crashCounts.R
\name{crashCounts}
\alias{crashCounts}
\title{Merge and Aggregate Crash Data by Road Segments}
\usage{
crashCounts(
  roads,
  crashes,
  road_id_vars,
  start_mp,
  end_mp,
  crash_mp,
  conditions = NULL,
  countvarname = "Total_Crashes"
)
}
\arguments{
\item{roads}{A dataframe containing road segment data. Expected fields include segment identifiers
specified by `road_id_vars`, as well as the starting and ending milepost variables specified
by `start_mp` and `end_mp`.}

\item{crashes}{A dataframe containing crash data. Expected fields include crash location specified
by `crash_mp` and the same segment identifiers as `roads` specified by `road_id_vars`.}

\item{road_id_vars}{A character vector specifying the column names that uniquely identify each road
segment. For example, `c("County", "RouteNo", "Region")` or `c("RouteID")` if only a single
identifier is required. The values in `road_id_vars` must be strings (characters).}

\item{start_mp}{A string specifying the name of the column in `roads` that represents the starting
milepost or measurement for each road segment. The values in `start_mp` must be numeric.}

\item{end_mp}{A string specifying the name of the column in `roads` that represents the ending
milepost or measurement for each road segment. The values in `end_mp` must be numeric.}

\item{crash_mp}{A string specifying the name of the column in `crashes` that represents the crash
location (e.g., milepoint or distance). The values in `crash_mp` must be numeric.}

\item{conditions}{Optional. A condition or set of conditions for filtering the crashes dataframe
before aggregation. This can be specified as a character string (e.g., `"Severity > 2"`),
an expression (e.g., `expression(Severity > 2)`), or a list of expressions
(e.g., `list(expression(Severity > 2), expression(Time < 6 | Time > 18))`). If `NULL`
(default), no additional conditions are applied.}

\item{countvarname}{Optional. A string for the name of the new variable with the crash counts.
The default is "Total_Crashes" but should be specified by the user for any time conditions
are provided.}
}
\value{
A dataframe with the original road dataframe and the new variable that has the crash
        counts for the specified conditions (if any).
}
\description{
This function merges a roads dataset with a crashes dataset based on specified route identifiers
and segment boundaries. It then aggregates crash data per road segment per year, allowing
for optional user-specified conditions to filter crashes based on various attributes.
}
