% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intersections}
\alias{intersections}
\title{Sample Intersection Data}
\format{
An \code{sf} data frame with spatial point geometry and the following variables:
\describe{
  \item{Intersection_ID}{Intersection ID}
  \item{Lat_int}{Intersection Latitude}
  \item{Long_int}{Intersection Longitude}
  \item{major_AADT}{Major Road Annual Average Daily Traffic}
  \item{minor_AADT}{Minor Road Annual Average Daily Traffic}
}
}
\usage{
intersections
}
\description{
A sample data set containing intersection information for use in demonstrating the crash analysis functions.
}
\keyword{datasets}
